Scan added checkpoint and VIF checking fixed

+++ src/Actions/ComputeMembers/Scan.php
@@ -22,6 +22,8 @@ class Scan extends AbstractAction
 
     public function __construct(ComputeMembers $computeMember)
     {
+        UserHelper::setAdminAsCurrentUser();
+
         $this->model = $computeMember;
 
         $this->queue = 'iaas';
@@ -31,39 +33,57 @@ class Scan extends AbstractAction
 
     public function handle()
     {
-        UserHelper::setAdminAsCurrentUser();
-
         $this->setProgress(0, 'Initiate compute member started');
 
-        $this->setProgress(10, 'Updating compute member information');
-        ComputeMemberXenService::updateMemberInformation($this->model);
+        if($this->shouldRunCheckpoint(10)) {
+            ComputeMemberXenService::updateMemberInformation($this->model);
+            $this->setProgress(10, 'Updating compute member information');
+        }
 
-        $this->setProgress(15, 'Removing vlans which are deleted from compute member');
-        ComputeMemberXenService::removeDeletedVlans($this->model);
+        if($this->shouldRunCheckpoint(15)) {
+            ComputeMemberXenService::removeDeletedVlans($this->model);
+            $this->setProgress(15, 'Removing vlans which are deleted from compute member');
+        }
 
-        $this->setProgress(20, 'Updating compute member network interface information');
-        ComputeMemberXenService::updateInterfaceInformation($this->model);
+        if($this->shouldRunCheckpoint(20)) {
+            ComputeMemberXenService::updateInterfaceInformation($this->model);
+            $this->setProgress(20, 'Updating compute member network interface information');
+        }
 
-        $this->setProgress(30, 'Updating compute member storage repository information');
-        ComputeMemberXenService::updateMissingVlans($this->model);
+        if($this->shouldRunCheckpoint(30)) {
+            ComputeMemberXenService::updateMissingVlans($this->model);
+            $this->setProgress(30, 'Updating compute member storage repository information');
+        }
 
-        $this->setProgress(40, 'Updating compute member bridges/networks information');
-        ComputeMemberXenService::updateNetworkInformation($this->model);
+        if($this->shouldRunCheckpoint(40)) {
+            ComputeMemberXenService::updateNetworkInformation($this->model);
+            $this->setProgress(40, 'Updating compute member bridges/networks information');
+        }
 
-        $this->setProgress(60, 'Updating compute member storage volume information');
-        ComputeMemberXenService::updateStorageVolumes($this->model);
+        if($this->shouldRunCheckpoint(60)) {
+            ComputeMemberXenService::updateStorageVolumes($this->model);
+            $this->setProgress(60, 'Updating compute member storage volume information');
+        }
 
-        $this->setProgress(65, 'Updating VMs in compute member:');
-        (new \NextDeveloper\IAAS\Actions\ComputeMembers\ScanVirtualMachines($this->model))->handle();
+        if($this->shouldRunCheckpoint(65)) {
+            (new \NextDeveloper\IAAS\Actions\ComputeMembers\ScanVirtualMachines($this->model))->handle();
+            $this->setProgress(65, 'Updating VMs in compute member:');
+        }
 
-        $this->setProgress(70, 'Updating compute member resources');
-        ComputeMemberXenService::updateMemberInformation($this->model);
+        if($this->shouldRunCheckpoint(70)) {
+            ComputeMemberXenService::updateMemberInformation($this->model);
+            $this->setProgress(70, 'Updating compute member resources');
+        }
 
-        $this->setProgress(80, 'Updating network information');
-        ComputeMemberXenService::updateConnectionInformation($this->model);
+        if($this->shouldRunCheckpoint(80)) {
+            ComputeMemberXenService::updateConnectionInformation($this->model);
+            $this->setProgress(80, 'Updating network information');
+        }
 
-        $this->setProgress(90, 'Creating network member');
-        NetworkMemberXenService::createNetworkMemberFromComputeMember($this->model);
+        if($this->shouldRunCheckpoint(90)) {
+            NetworkMemberXenService::createNetworkMemberFromComputeMember($this->model);
+            $this->setProgress(90, 'Creating network member');
+        }
 
         Events::fire('scanned:NextDeveloper\IAAS\ComputeMembers', $this->model);
